+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   * Introduced "external descriptors", new IDL syntax for declaring "external" types. This syntax
     replaces `@Cpp(External*)` group of IDL attributes.
   * Added support for "external" structs and enums in Java, Swift, and Dart.
+### Deprecated:
+  * `@Java(Builder)` IDL attribute is now deprecated. It still works the same way as before but
+    its use is discouraged.
 
 ## 7.1.6
 Release date: 2020-07-09